+++ unsorted_include_todo/IntroModeState.h
@@ -11,17 +11,17 @@
  */
 
 struct ModeState {
-	virtual void update(u32&);          // _08
-	virtual void postRender(Graphics&); // _0C
-	virtual void postUpdate();          // _10
+	virtual void update(unsigned long&); // _08
+	virtual void postRender(Graphics&);  // _0C
+	virtual void postUpdate();           // _10
 };
 
 /**
  * @brief TODO
  */
 struct IntroModeState : public ModeState {
-	virtual void update(u32&);          // _08
-	virtual void postRender(Graphics&); // _0C
+	virtual void update(unsigned long&); // _08
+	virtual void postRender(Graphics&);  // _0C
 };
 
 #endif
