+++ unsorted_include_todo/ModeState.h
@@ -14,9 +14,9 @@
  * @brief TODO
  */
 struct ModeState {
-	virtual void update(u32&);          // _08
-	virtual void postRender(Graphics&); // _0C
-	virtual void postUpdate();          // _10
+	virtual void update(unsigned long&); // _08
+	virtual void postRender(Graphics&);  // _0C
+	virtual void postUpdate();           // _10
 };
 
 #endif
